requirements.txt
PyQt5>=5.15.9
transformers>=4.36.0
torch>=2.1.0
python-docx>=1.1.0
PyMuPDF>=1.23.3
nltk>=3.8.1
